+++ mdtraj/formats/pdb/pdbfile.py
@@ -51,9 +51,8 @@ import numpy as np
 import xml.etree.ElementTree as etree
 from copy import copy
 from mdtraj.formats.pdb.pdbstructure import PdbStructure
-from mdtraj.io import open_maybe_zipped
 from mdtraj.core.topology import Topology
-from mdtraj.utils import ilen, cast_indices, in_units_of
+from mdtraj.utils import ilen, cast_indices, in_units_of, open_maybe_zipped
 from mdtraj.formats.registry import _FormatRegistry
 from mdtraj.core import element as elem
 from mdtraj.utils import six
